backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==1.0.0
Django==4.0.6
django-anymail==8.6
django-appconf==1.0.5
django-cloudinary-storage==0.3.0
django-filter==22.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django-user-accounts==3.2.0
djangorestframework==3.13.1
exif==1.3.5
ExifRead==3.0.0
gunicorn==20.1.0
hachoir==3.1.3
idna==3.3
jsonpickle==2.2.0
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
plum-py==0.8.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.27
pytz==2022.1
requests==2.28.1
sentry-sdk==1.9.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
